+++ src/Dates/CalendarId.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace ICanBoogie\CLDR\Dates;
+
+enum CalendarId : string
+{
+    case GREGORIAN = 'gregorian';
+    case BUDDHIST = 'buddhist';
+    case CHINESE = 'chinese';
+    case COPTIC = 'coptic';
+    case DANGI = 'dangi';
+    case ETHIOPIC = 'ethiopic';
+    case HEBREW = 'hebrew';
+    case INDIAN = 'indian';
+    case ISLAMIC = 'islamic';
+    case JAPANESE = 'japanese';
+    case PERSIAN = 'persian';
+    case ROC = 'roc';
+}